refactor(metrics): extract useExploreMetric hook from ExploreMetricButton
